adguardian

Module welcome

Structs

Crate 🔒

CratesIoResponse 🔒

Functions

check_for_updates 🔒

Checks for updates to the crate, and prints a message if an update is available

check_version 🔒

Given a possibly undefined version number, check if it’s present and supported

get_env 🔒

Given a key, get the value from the environmental variables, and print it to the console

get_latest_version 🔒

Gets the latest version of the crate from crates.io

print_ascii_art 🔒

Prints the AdGuardian ASCII art to console

print_error 🔒

Print error message, along with (optional) stack trace, then exit

print_info 🔒

Reusable function that just prints success messages to the console

verify_connection 🔒

With the users specified AdGuard details, verify the connection (exit on fail)

welcome

Initiate the welcome script This function will: